Bug Fix: 1323204
The root cause of this issue is that libvirt can not free up vhost-net for the tap device
which has been already deleted by nova compute (contrail) vif driver. As a result, qemu
cores when interface-detach is attempted. Since libvirt doesn't create the tap device (for
ethernet type interface), it can not delete it as it doens't own the device. The fix is to
not delete the tap device from unplug call, rather delegate this work to a deferred thread,
which deletes it after vhost-net is freed up by libvirt.

Change-Id: I4588f571b6699a64a931b492debd8dee14bf4b3f
Signed-off-by: Anirban Chakraborty <abchak@juniper.net>