Build skeleton for 'vrouter' os-vif plugin
The current os-vif plugin operates on the contrail_vrouter VIF type.
This means that the only plugging mode supported is the vhost-user
method with the DPDK datapath.

This changeset renames the 'contrail_vrouter' plugin back to 'vrouter'
and adds a small shim for a 'contrail_vrouter' plugin. The
'contrail_vrouter' plugin will be a strict subset of the 'vrouter'
plugin, supporting only the vhost-user datapath for backwards
compatibility.

This changeset clones the existing unit tests to the 'contrail_vrouter'
plugin, ensuring that the interfaces can be maintained as work progresses
on the 'vrouter' plugin.

This changeset also cleans up a number of PEP8 issues and
failing unit tests in the current os-vif plugin:

    * Whitespace issues
    * Import ordering
    * Filling in proper UUID's in tests
    * Better handling of exceptions when config is not available

Change-Id: Ica76793e7977616236b07efbfaa115fe85c97fab
Partial-Bug: #1773981
Signed-off-by: Jan Gutter <jan.gutter@netronome.com>
Acked-by: Frikkie Scholtz <frikkie.scholtz@netronome.com>