Refactor call to vrouter-port-control
This changeset breaks the vrouter-port-control call out into
discrete functions. This is a cleanup, preserving existing
test expectations, to allow for common code to be used with multiple
plugging modes.

The unit tests have been updated to reflect that the calls to the
vrouter-port-control script are now discrete parameters rather than
a single string.

Change-Id: Ia554b5d1e3036eff06762dd4e8fe71a9e6715a35
Partial-Bug: #1773981
Signed-off-by: Jan Gutter <jan.gutter@netronome.com>