Add kernel vrouter.ko datapath plugging support
This change adds the missing sections to support kernel TAP device
plugging for the os-vif plugin. Currently this is keyed off the VIF type
passed to the plugin.

Note: once upstream os-vif gains support for multiqueue VIFGeneric, this
os-vif plugin should be a complete drop-in replacement for the Nova
legacy VIF.

* Unit tests have been updated:
  contrail_vrouter os-vif plugin tests only vhostuser plugging
  vrouter os-vif plugin tests both supported methods

Change-Id: I74dca19a6c8a5676caa6423619b1f593a8cc1b3e
Partial-Bug: #1773981
Signed-off-by: Jan Gutter <jan.gutter@netronome.com>